requirements.txt
blinker==1.8.2
boto3==1.35.71
botocore==1.35.71
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
Flask==3.0.3
flask-marshmallow==1.2.1
fonttools==4.55.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.7
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.3
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
s3transfer==0.10.4
setuptools==75.1.0
six==1.16.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.44.0
